feat(admin): full-width cash chart with net burn and runway beneath
Drops the Cash and Net burn stat cards — both duplicated their charts.
Cash trajectory goes full width, with the net burn chart and a runway
card (carrying the cash ÷ burn inputs in its subtitle) side by side
underneath.
