--- apps/admin/src/app/(dashboard)/components/CashRunwayTile/index.ts
@@ -1,1 +1,0 @@
-export { CashRunwayTile } from "./CashRunwayTile";
