+++ apps/admin/src/app/(dashboard)/components/RunwayTile/index.ts
@@ -1,0 +1,1 @@
+export { RunwayTile } from "./RunwayTile";
